changed project/folder hierarchy and naming
