+++ Samples/NetworkMediaStreamer/NetworkMediaStreamerBase/inc/NwMediaStreamSinkBase.h
